refactor:step 3 Refactoring
