+++ src/js/index.js
@@ -1,5 +1,6 @@
 import { $ } from "./utils/dom.js";
 import { store } from "./store/index.js";
+import MenuApi from "../api/index.js";
 
 // step 3 요구사항
 // TODO 서버 요청 부분
@@ -11,8 +12,8 @@ import { store } from "./store/index.js";
 // [x] - 서버에 메뉴 삭제를 요청한다.
 
 // TODO 리팩터링 부분
-// [] - localStorage에 저장하는 로직은 지운다.
-// [] - fetch 비동기 api를 사용하는 부분을 async await을 사용하여 구현한다.
+// [x] - localStorage에 저장하는 로직은 지운다.
+// [x] - fetch 비동기 api를 사용하는 부분을 async await을 사용하여 구현한다.
 
 // TODO 사용자 설정
 // [] - API 통신이 실패하는 경우에 대해 사용자가 알 수 있게 alert으로 예외처리를 진행한다.
@@ -20,61 +21,6 @@ import { store } from "./store/index.js";
 // API 통신 방법 fetch('url', option)
 
 const BASE_URL = "http://localhost:3000/api";
-const MenuApi = {
-  async getAllMenuByCategory(category) {
-    const response = await fetch(`${BASE_URL}/category/${category}/menu`);
-    return response.json();
-  },
-  async createMenu(category, name) {
-    const response = await fetch(`${BASE_URL}/category/${category}/menu`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ name }),
-    });
-
-    if (!response.ok) {
-      console.error("에러가 발생했습니다.");
-    }
-  },
-  async editMenuName(category, menuId, name) {
-    const response = await fetch(
-      `${BASE_URL}/category/${category}/menu/${menuId}`,
-      {
-        method: "PUT",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ name }),
-      }
-    );
-    if (!response.ok) {
-      console.error("에러가 발생했습니다.");
-    }
-    return response.json();
-  },
-  async toggleSoldOutMenu(category, menuId) {
-    const response = await fetch(
-      `${BASE_URL}/category/${category}/menu/${menuId}/soldout`,
-      {
-        method: "PUT",
-      }
-    );
-    if (!response.ok) {
-      console.error("에러가 발생했습니다.");
-    }
-  },
-  async removeMenu(category, menuId) {
-    const response = await fetch(
-      `${BASE_URL}/category/${category}/menu/${menuId}`,
-      {
-        method: "DELETE",
-      }
-    );
-    if (!response.ok) {
-      console.error("에러가 발생했습니다.");
-    }
-  },
-};
 function App() {
   this.menu = {
     espresso: [],
@@ -93,7 +39,10 @@ function App() {
     initEventListener();
   };
 
-  const render = () => {
+  const render = async () => {
+    this.menu[this.currentCategory] = await MenuApi.getAllMenuByCategory(
+      this.currentCategory
+    );
     const template = this.menu[this.currentCategory]
       .map((item) => {
         return `
@@ -139,9 +88,6 @@ function App() {
     const menuName = $("#menu-name").value;
     // 메뉴 추가하는 api 요청
     await MenuApi.createMenu(this.currentCategory, menuName);
-    this.menu[this.currentCategory] = await MenuApi.getAllMenuByCategory(
-      this.currentCategory
-    );
     render();
     $("#menu-name").value = "";
   };
@@ -151,9 +97,6 @@ function App() {
     const $MenuName = e.target.closest("li").querySelector(".menu-name");
     const updateMenuName = prompt("메뉴를 수정하세요", $MenuName.innerText);
     await MenuApi.editMenuName(this.currentCategory, menuId, updateMenuName);
-    this.menu[this.currentCategory] = await MenuApi.getAllMenuByCategory(
-      this.currentCategory
-    );
     render();
   };
 
@@ -161,19 +104,23 @@ function App() {
     if (confirm("정말 삭제하시겠습니까?")) {
       const menuId = e.target.closest("li").dataset.menuId;
       await MenuApi.removeMenu(this.currentCategory, menuId);
-      this.menu[this.currentCategory] = await MenuApi.getAllMenuByCategory(
-        this.currentCategory
-      );
       render();
     }
   };
   const soldOutMenu = async (e) => {
     const menuId = e.target.closest("li").dataset.menuId;
     await MenuApi.toggleSoldOutMenu(this.currentCategory, menuId);
-    this.menu[this.currentCategory] = await MenuApi.getAllMenuByCategory(
-      this.currentCategory
-    );
     render();
+  };
+
+  const changeCategory = (e) => {
+    const isCategoryButton = e.target.classList.contains("cafe-category-name");
+    if (isCategoryButton) {
+      const CategoryName = e.target.dataset.categoryName;
+      this.currentCategory = CategoryName;
+      $("#category-title").innerText = `${e.target.innerText} 메뉴 관리`;
+      render();
+    }
   };
   const initEventListener = () => {
     // form 태그 자동 전송 막기
@@ -205,19 +152,7 @@ function App() {
       }
     });
 
-    $("nav").addEventListener("click", async (e) => {
-      const isCategoryButton =
-        e.target.classList.contains("cafe-category-name");
-      if (isCategoryButton) {
-        const CategoryName = e.target.dataset.categoryName;
-        this.currentCategory = CategoryName;
-        $("#category-title").innerText = `${e.target.innerText} 메뉴 관리`;
-        this.menu[this.currentCategory] = await MenuApi.getAllMenuByCategory(
-          this.currentCategory
-        ); //data 뒤에 것이 알아서 카멜케이스로 바뀜
-        render();
-      }
-    });
+    $("nav").addEventListener("click", changeCategory);
   };
 }
